divide off shapes from element_commands
